+++ framework/Source/CPTXYPlotSpace.h
@@ -3,6 +3,13 @@
 
 @class CPTPlotRange;
 
+typedef enum _CPTRubberBandType {
+    kCPTRubberBandNone,
+    kCPTRubberBandLeft,
+    kCPTRubberBandRight
+}
+CPTRubberBandType;
+
 @interface CPTXYPlotSpace : CPTPlotSpace {
     @private
     CPTPlotRange *xRange;
@@ -13,6 +20,9 @@
     CPTScaleType yScaleType;
     CGPoint lastDragPoint;
     BOOL isDragging;
+    BOOL hasMomentum;
+    CPTRubberBandType rubberBand;
+    CGPoint momentum;
 }
 
 @property (nonatomic, readwrite, copy) CPTPlotRange *xRange;
@@ -21,5 +31,9 @@
 @property (nonatomic, readwrite, copy) CPTPlotRange *globalYRange;
 @property (nonatomic, readwrite, assign) CPTScaleType xScaleType;
 @property (nonatomic, readwrite, assign) CPTScaleType yScaleType;
+
+@property (nonatomic, readwrite) BOOL allowsMomentum;
+@property (nonatomic, readwrite) BOOL elasticGlobalXRange;
+@property (nonatomic, readwrite) BOOL elasticGlobalYRange;
 
 @end